Improved OTRS installation.

+++ otrs/functions.sh
@@ -20,6 +20,8 @@
 . ./util_functions.sh
 
 #Default configuration values
+DEFAULT_OTRS_DB_HOST="mariadb"
+DEFAULT_OTRS_DB_PORT="3306"
 DEFAULT_OTRS_ADMIN_EMAIL="admin@example.com"
 DEFAULT_OTRS_ORGANIZATION="Example Company"
 DEFAULT_OTRS_SYSTEM_ID="98"
@@ -32,18 +34,23 @@ DEFAULT_OTRS_CUSTOMER_LOGO_RIGHT="25"
 DEFAULT_OTRS_CUSTOMER_LOGO_TOP="2"
 DEFAULT_OTRS_CUSTOMER_LOGO_WIDTH="135"
 OTRS_BACKUP_DIR="/var/otrs/backups"
-OTRS_CONFIG_DIR="${OTRS_ROOT}Kernel"
-OTRS_CONFIG_MOUNT_DIR="/config/"
+OTRS_CONFIG_DIR="${OTRS_ROOT}/Kernel"
+OTRS_CONFIG_MOUNT_DIR="/config"
 
 [ -z "${OTRS_INSTALL}" ] && OTRS_INSTALL="no"
 
-mysqlcmd="mysql -uroot -h mariadb -p$MYSQL_ROOT_PASSWORD "
+[ -z "${OTRS_DB_HOST}" ] && OTRS_DB_HOST="$DEFAULT_OTRS_DB_HOST"
+[ -z "${OTRS_DB_PORT}" ] && OTRS_DB_PORT="$DEFAULT_OTRS_DB_PORT"
 
-function create_db(){
+function mysqlcmd() {
+    mysql -uroot --protocol=TCP --host="${OTRS_DB_HOST}" --port="${OTRS_DB_PORT}" --password="${MYSQL_ROOT_PASSWORD}" "$@"
+}
+
+function create_db() {
   print_info "Creating OTRS database..."
-  $mysqlcmd -e "CREATE DATABASE IF NOT EXISTS otrs;"
+  mysqlcmd -e "CREATE DATABASE IF NOT EXISTS otrs;"
   [ $? -gt 0 ] && print_error "Couldn't create OTRS database !!" && exit 1
-  $mysqlcmd -e " GRANT ALL ON otrs.* to 'otrs'@'%' identified by '$OTRS_DB_PASSWORD'";
+  mysqlcmd -e " GRANT ALL ON otrs.* to 'otrs'@'%' identified by '$OTRS_DB_PASSWORD'";
   [ $? -gt 0 ] && print_error "Couldn't create database user !!" && exit 1
 }
 
@@ -54,11 +61,11 @@ function restore_backup(){
 
   #As this is a restore, drop database first.
 
-  $mysqlcmd -e 'use otrs'
+  mysqlcmd -e 'use otrs'
   if [ $? -eq 0  ]; then
     if [ "$OTRS_DROP_DATABASE" == "yes" ]; then
       print_info "OTRS_DROP_DATABASE=\e[92m$OTRS_DROP_DATABASE\e[0m, Dropping existing database\n"
-      $mysqlcmd -e 'drop database otrs'
+      mysqlcmd -e 'drop database otrs'
     else
       print_error "Couldn't load OTRS backup, databse already exists !!" && exit 1
     fi
@@ -67,58 +74,58 @@ function restore_backup(){
   create_db
   update_config_password $OTRS_DB_PASSWORD
 
-  #Make a copy of installed skins so they aren't overwritten by the backup.
+  # Make a copy of installed skins so they aren't overwritten by the backup.
   tmpdir=`mktemp -d`
-  [ ! -z $OTRS_AGENT_SKIN ] && cp -rp ${SKINS_PATH}Agent $tmpdir/
-  [ ! -z $OTRS_CUSTOMER_SKIN ] && cp -rp ${SKINS_PATH}Customer $tmpdir/
-  #Run restore backup command
-  ${OTRS_ROOT}scripts/restore.pl -b $OTRS_BACKUP_DIR/$1 -d ${OTRS_ROOT}
+  [ ! -z "$OTRS_AGENT_SKIN" ] && cp -rp ${SKINS_PATH}Agent $tmpdir/
+  [ ! -z "$OTRS_CUSTOMER_SKIN" ] && cp -rp ${SKINS_PATH}Customer $tmpdir/
+  # Run restore backup command
+  ${OTRS_ROOT}/scripts/restore.pl -b $OTRS_BACKUP_DIR/$1 -d ${OTRS_ROOT}/
   [ $? -gt 0 ] && print_error "Couldn't load OTRS backup !!" && exit 1
 
   backup_version=`tar -xOf $OTRS_BACKUP_DIR/$1/Application.tar.gz ./RELEASE|grep -o 'VERSION = [^,]*' | cut -d '=' -f2 |tr -d '[[:space:]]'`
   OTRS_INSTALLED_VERSION=`echo $OTRS_VERSION|cut -d '-' -f1`
-  print_warning "OTRS version of backup being restored: \e[1;31m$backup_version\e[1;0m"
-  print_warning "OTRS version of this container: \e[1;31m$OTRS_INSTALLED_VERSION\e[1;0m"
+  print_warning "OTRS version of backup being restored: \e[1;31m${backup_version}\e[1;0m"
+  print_warning "OTRS version of this container: \e[1;31m${OTRS_INSTALLED_VERSION}\e[1;0m"
 
-  check_version $OTRS_INSTALLED_VERSION $backup_version
+  check_version "$OTRS_INSTALLED_VERSION" "$backup_version"
   if [ $? -eq 0 ]; then
     print_warning "Backup version older than current OTRS version, fixing..."
-    #Update version on ${OTRS_ROOT}/RELEASE so it the website shows the correct version.
-    sed -i -r "s/(VERSION *= *).*/\1$OTRS_INSTALLED_VERSION/" ${OTRS_ROOT}RELEASE
+    # Update version on ${OTRS_ROOT}/RELEASE so it the website shows the correct version.
+    sed -i -r "s/(VERSION *= *).*/\1$OTRS_INSTALLED_VERSION/" ${OTRS_ROOT}/RELEASE
     print_info "Done."
   fi
 
-  #Restore configured password overwritten by restore
-  update_config_password $OTRS_DB_PASSWORD
-  #Copy back skins over restored files
-  [ ! -z $OTRS_CUSTOMER_SKIN ] && cp -rfp $tmpdir/* ${SKINS_PATH} && rm -fr $tmpdir
+  # Restore configured password overwritten by restore
+  update_config_password "$OTRS_DB_PASSWORD"
+  # Copy back skins over restored files
+  [ ! -z "$OTRS_CUSTOMER_SKIN" ] && cp -rfp $tmpdir/* ${SKINS_PATH} && rm -fr $tmpdir
 
   #Update the skin preferences  in the users from the backup
   set_users_skin
 }
 
 # return 0 if program version is equal or greater than check version
-check_version()
-{
+check_version() {
     local version=$1 check=$2
     local winner=$(echo -e "$version\n$check" | sed '/^$/d' | sort -nr | head -1)
     [[ "$winner" = "$version" ]] && return 0
     return 1
 }
 
-function random_string(){
+function random_string() {
   echo `cat /dev/urandom | tr -dc 'a-zA-Z0-9' | fold -w 8 | head -n 1`
 }
 
-function update_config_password(){
+function update_config_password() {
   #Change database password on configuration file
-  sed  -i "s/\($Self->{'DatabasePw'} *= *\).*/\1'$1';/" ${OTRS_ROOT}Kernel/Config.pm
+  sed  -i "s/\($Self->{DatabasePw} *= *\).*/\1'$1';/" ${OTRS_ROOT}/Kernel/Config.pm
 }
 
-function copy_default_config(){
-  print_info "Copying configuration file..."
-  cp -f ${OTRS_ROOT}docker/defaults/Config.pm.default ${OTRS_ROOT}Kernel/Config.pm
-  [ $? -gt 0 ] && print_error "\n\e[1;31mERROR:\e[0m Couldn't load OTRS config file !!\n" && exit 1
+function copy_default_config() {
+  #print_info "Copying configuration file..."
+  #cp -f "${OTRS_ROOT}/docker/defaults/Config.pm.default" "${OTRS_ROOT}/Kernel/Config.pm"
+  #[ $? -gt 0 ] && print_error "\n\e[1;31mERROR:\e[0m Couldn't load OTRS config file !!\n" && exit 1
+    :
 }
 
 function set_variables(){
@@ -148,73 +155,81 @@ function set_variables(){
   fi
 }
 
-function load_defaults(){
-  set_variables
-  #Check if a host-mounted volume for configuration storage was added to this
-  #container
-  check_host_mount_dir
-  copy_default_config
-  update_config_password $OTRS_DB_PASSWORD
+function load_defaults() {
+    set_variables
+    # Check if a host-mounted volume for configuration storage was added to this
+    # container
+    check_host_mount_dir
+    copy_default_config
+    update_config_password "$OTRS_DB_PASSWORD"
 
-  #Add default config options
-  sed -i "/$Self->{'SecureMode'} = 1;/a \
- \$Self->{'FQDN'} = '$OTRS_HOSTNAME';\
+    # Add default config options
+    sed -i "/#[[:space:]]*\$DIBI\$[[:space:]]*$/a\\
+\n\$Self->{'FQDN'} = '$OTRS_HOSTNAME';\
 \n\$Self->{'AdminEmail'} = '$OTRS_ADMIN_EMAIL';\
+\n\$Self->{'DatabaseHost'} = '$OTRS_DB_HOST';\
 \n\$Self->{'Organization'} = '$OTRS_ORGANIZATION';\
 \n\$Self->{'CustomerHeadline'} = '$OTRS_ORGANIZATION';\
 \n\$Self->{'SystemID'} = '$OTRS_SYSTEM_ID';\
-\n\$Self->{'PostMaster::PreFilterModule::NewTicketReject::Sender'} = 'noreply@${OTRS_HOSTNAME}';"\
-\n\$Self->{'PostmasterFollowUpSearchInRaw'} = '1';"\
-\n\$Self->{'PostmasterFollowUpSearchInBody'} = '1';"\
-\n\$Self->{'PostmasterFollowUpSearchInAttachment'} = '1';"\
-\n\$Self->{'PostmasterFollowUpSearchInReferences'} = '1';"\
- ${OTRS_ROOT}Kernel/Config.pm
+\n\$Self->{'PostMaster::PreFilterModule::NewTicketReject::Sender'} = 'noreply@${OTRS_HOSTNAME}';\
+\n\$Self->{'PostmasterFollowUpSearchInRaw'} = '1';\
+\n\$Self->{'PostmasterFollowUpSearchInBody'} = '1';\
+\n\$Self->{'PostmasterFollowUpSearchInAttachment'} = '1';\
+\n\$Self->{'PostmasterFollowUpSearchInReferences'} = '1';\
+\n\$Self->{DatabaseDSN} = \"DBI:mysql:database=\$Self->{Database};host=\$Self->{DatabaseHost};\";"\
+        ${OTRS_ROOT}/Kernel/Config.pm
 
-  #Check if database doesn't exists yet (it could if this is a container redeploy)
-  $mysqlcmd -e 'use otrs'
-  if [ $? -gt 0 ]; then
-    create_db
+    # Check if database doesn't exists yet (it could if this is a container redeploy)
+    if ! mysqlcmd -e 'use otrs'; then
+        create_db
 
-    #Check that a backup isn't being restored
-    if [ "$OTRS_INSTALL" == "no" ]; then
-      print_info "Loading default db schema..."
-      $mysqlcmd otrs < ${OTRS_ROOT}scripts/database/otrs-schema.mysql.sql
-      [ $? -gt 0 ] && print_error "\n\e[1;31mERROR:\e[0m Couldn't load OTRS database schema !!\n" && exit 1
-      print_info "Loading initial db inserts..."
-      $mysqlcmd otrs < ${OTRS_ROOT}scripts/database/otrs-initial_insert.mysql.sql
-      [ $? -gt 0 ] && print_error "\n\e[1;31mERROR:\e[0m Couldn't load OTRS database initial inserts !!\n" && exit 1
+        # Check that a backup isn't being restored
+        if [ "$OTRS_INSTALL" == "no" ]; then
+
+            print_info "Loading default db schema..."
+            mysqlcmd otrs < "${OTRS_ROOT}/scripts/database/otrs-schema.mysql.sql"
+            [ $? -ne 0 ] && print_error "\n\e[1;31mERROR:\e[0m Couldn't load OTRS database schema !!\n" && exit 1
+            
+            print_info "Loading initial db inserts..."
+            mysqlcmd otrs < "${OTRS_ROOT}/scripts/database/otrs-initial_insert.mysql.sql"
+            [ $? -ne 0 ] && print_error "\n\e[1;31mERROR:\e[0m Couldn't load OTRS database initial inserts !!\n" && exit 1
+
+            print_info "Loading db post initialization..."
+            mysqlcmd otrs < "${OTRS_ROOT}/scripts/database/otrs-schema-post.mysql.sql"
+            [ $? -ne 0 ] && print_error "\n\e[1;31mERROR:\e[0m Couldn't load OTRS database post initialization !!\n" && exit 1
+
+        fi
+    else
+        print_warning "otrs database already exists, Ok."
     fi
-  else
-    print_warning "otrs database already exists, Ok."
-  fi
 }
 
-function set_default_language(){
-  if [ ! -z $OTRS_LANGUAGE ]; then
-    print_info "Setting default language to: \e[92m'$OTRS_LANGUAGE'\e[0m"
-    sed -i "/$Self->{'SecureMode'} = 1;/a \
-    \$Self->{'DefaultLanguage'} = '$OTRS_LANGUAGE';"\
-    ${OTRS_ROOT}Kernel/Config.pm
- fi
+function set_default_language() {
+    if [ ! -z $OTRS_LANGUAGE ]; then
+        print_info "Setting default language to: \e[92m'$OTRS_LANGUAGE'\e[0m"
+        sed -i "/#[[:space:]]*\$DIBI\$[[:space:]]*$/a\\
+    \n\$Self->{'DefaultLanguage'} = '$OTRS_LANGUAGE';"\
+            ${OTRS_ROOT}/Kernel/Config.pm
+    fi
 }
 
 function set_ticker_counter() {
   if [ ! -z "${OTRS_TICKET_COUNTER}" ]; then
     print_info "Setting the start of the ticket counter to: \e[92m'$OTRS_TICKET_COUNTER'\e[0m"
-    echo "$OTRS_TICKET_COUNTER" > ${OTRS_ROOT}var/log/TicketCounter.log
+    echo "$OTRS_TICKET_COUNTER" > ${OTRS_ROOT}/var/log/TicketCounter.log
   fi
   if [ ! -z $OTRS_NUMBER_GENERATOR ]; then
     print_info "Setting ticket number generator to: \e[92m'$OTRS_NUMBER_GENERATOR'\e[0m"
-    sed -i "/$Self->{'SecureMode'} = 1;/a \$Self->{'Ticket::NumberGenerator'} =  'Kernel::System::Ticket::Number::${OTRS_NUMBER_GENERATOR}';"\
-     ${OTRS_ROOT}Kernel/Config.pm
+    sed -i "/#[[:space:]]*\$DIBI\$[[:space:]]*$/a \$Self->{'Ticket::NumberGenerator'} =  'Kernel::System::Ticket::Number::${OTRS_NUMBER_GENERATOR}';"\
+     ${OTRS_ROOT}/Kernel/Config.pm
   fi
 }
 
 function set_skins() {
-  [ ! -z $OTRS_AGENT_SKIN ] &&  sed -i "/$Self->{'SecureMode'} = 1;/a \
-\$Self->{'Loader::Agent::DefaultSelectedSkin'} =  '$OTRS_AGENT_SKIN';\
+  [ ! -z $OTRS_AGENT_SKIN ] &&  sed -i "/#[[:space:]]*\$DIBI\$[[:space:]]*$/a\\
+\n\$Self->{'Loader::Agent::DefaultSelectedSkin'} =  '$OTRS_AGENT_SKIN';\
 \n\$Self->{'Loader::Customer::SelectedSkin'} =  '$OTRS_CUSTOMER_SKIN';"\
- ${OTRS_ROOT}Kernel/Config.pm
+ ${OTRS_ROOT}/Kernel/Config.pm
 
   #Set Agent interface logo
   [ ! -z $OTRS_AGENT_LOGO ] && set_agent_logo
@@ -225,7 +240,7 @@ function set_skins() {
 
 function set_users_skin(){
   print_info "Updating default skin for users in backup..."
-  $mysqlcmd -e "UPDATE user_preferences SET preferences_value = '$OTRS_AGENT_SKIN' WHERE preferences_key = 'UserSkin'" otrs
+  mysqlcmd -e "UPDATE user_preferences SET preferences_value = '$OTRS_AGENT_SKIN' WHERE preferences_key = 'UserSkin'" otrs
   [ $? -gt 0 ] && print_error "\n\e[1;31mERROR:\e[0m Couldn't change default skin for existing users !!\n"
 }
 
@@ -245,16 +260,16 @@ function set_logo () {
   logo_width=$5
   logo_url=$6
 
-  sed -i "/$Self->{'SecureMode'} = 1;/a \
- \$Self->{'${interface}Logo'} =  {\n'StyleHeight' => '${logo_height}px',\
+  sed -i "/#[[:space:]]*\$DIBI\$[[:space:]]*$/a\\
+\n\$Self->{'${interface}Logo'} =  {\n'StyleHeight' => '${logo_height}px',\
 \n'StyleRight' => '${logo_right}px',\
 \n'StyleTop' => '${logo_top}px',\
 \n'StyleWidth' => '${logo_width}px',\
-\n'URL' => '$logo_url'\n};" ${OTRS_ROOT}Kernel/Config.pm
+\n'URL' => '$logo_url'\n};" ${OTRS_ROOT}/Kernel/Config.pm
 }
 
 # function set_customer_logo() {
-#   sed -i "/$Self->{'SecureMode'} = 1;/a\$Self->{'CustomerLogo'} =  {\n'StyleHeight' => '${OTRS_CUSTOMER_LOGO_HEIGHT}px',\n'StyleRight' => '${OTRS_CUSTOMER_LOGO_RIGHT}px',\n'StyleTop' => '${OTRS_CUSTOMER_LOGO_TOP}px',\n'StyleWidth' => '${OTRS_CUSTOMER_LOGO_WIDTH}px',\n'URL' => '$OTRS_CUSTOMER_LOGO'\n};" ${OTRS_ROOT}Kernel/Config.pm
+#   sed -i "/#[[:space:]]*\$DIBI\$[[:space:]]*$/a\ \$Self->{'CustomerLogo'} =  {\n'StyleHeight' => '${OTRS_CUSTOMER_LOGO_HEIGHT}px',\n'StyleRight' => '${OTRS_CUSTOMER_LOGO_RIGHT}px',\n'StyleTop' => '${OTRS_CUSTOMER_LOGO_TOP}px',\n'StyleWidth' => '${OTRS_CUSTOMER_LOGO_WIDTH}px',\n'URL' => '$OTRS_CUSTOMER_LOGO'\n};" ${OTRS_ROOT}/Kernel/Config.pm
 # }
 
 function set_fetch_email_time(){
@@ -267,7 +282,7 @@ function set_fetch_email_time(){
       sed -i -e '/otrs.PostMasterMailbox.pl/ s/^#*/#/' /var/spool/cron/otrs
     else
       #sed -i -e '/otrs.PostMasterMailbox.pl/ s/^#*//' /var/spool/cron/otrs
-      ${OTRS_ROOT}scripts/otrs_postmaster_time.sh $OTRS_POSTMASTER_FETCH_TIME
+      ${OTRS_ROOT}/scripts/otrs_postmaster_time.sh $OTRS_POSTMASTER_FETCH_TIME
     fi
   fi
 }
@@ -295,14 +310,19 @@ function check_host_mount_dir(){
       print_error "ERROR: Can't copy OTRS configuration to host-mounted volume ${OTRS_CONFIG_MOUNT_DIR}" && exit 1
     fi
   fi
-  print_info "Linking back \e[92m${OTRS_CONFIG_MOUNT_DIR}\e[0m to \e[92m${OTRS_CONFIG_DIR}\e[0m..."
+  
   #ln -s ${OTRS_CONFIG_MOUNT_DIR} ${OTRS_CONFIG_DIR}
-  mkdir -p ${OTRS_CONFIG_DIR}
-  cp -rp ${OTRS_CONFIG_MOUNT_DIR}/* ${OTRS_CONFIG_DIR}
-  if [ $? -eq 0 ];
-  then
-    print_info "Done."
+  mkdir -p "${OTRS_CONFIG_DIR}"
+  if [ -e "${OTRS_CONFIG_MOUNT_DIR}" ]; then
+      print_info "Linking back \e[92m${OTRS_CONFIG_MOUNT_DIR}\e[0m to \e[92m${OTRS_CONFIG_DIR}\e[0m..."
+      cp -rp ${OTRS_CONFIG_MOUNT_DIR}/* ${OTRS_CONFIG_DIR}
+      if [ $? -eq 0 ];
+      then
+          print_info "Done."
+      else
+          print_error "Can't create symlink to OTRS configuration on host-mounted volume ${OTRS_CONFIG_MOUNT_DIR}" && exit 1
+      fi
   else
-    print_error "Can't create symlink to OTRS configuration on host-mounted volume ${OTRS_CONFIG_MOUNT_DIR}" && exit 1
-fi
+      print_info "No configuration directory ${OTRS_CONFIG_MOUNT_DIR}"
+  fi
 }
